backend/requirements.txt
Django==5.2.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
Pillow==10.2.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
channels==4.0.0
daphne==4.0.0
stripe==7.9.0
python-dotenv==1.0.1
gunicorn==21.2.0
dj-database-url==2.1.0